refactor(coordinator): split orchestrator modules
